requirements.txt
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
crypto==1.4.1
cryptography==45.0.4
Deprecated==1.2.18
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.1
Flask-Bcrypt==1.0.1
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
Flask-WTF==1.2.2
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
Naked==0.1.32
ordered-set==4.1.0
packaging==25.0
pycparser==2.22
pycryptodome==3.23.0
Pygments==2.19.2
pymongo==4.13.2
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.4
rich==13.9.4
shellescape==3.8.1
SQLAlchemy==2.0.41
typing_extensions==4.14.0
urllib3==2.5.0
waitress==3.0.2
Werkzeug==3.1.3
wrapt==1.17.2
WTForms==3.2.1
